[glendid_users_17] Refactoriza instanciaciones y métodos repetidos Closes #17 y Closes #10
