+++ Makefile
@@ -3,6 +3,14 @@ install:
 		@pip install -r requirements.txt
 		@pip install -r test-requirements.txt
 
+#Deploy serverless app
+deploy:
+		@serverless deploy
+
+#Remove deployed apps
+remove:
+		@serverless remove
+
 #Create code lint report
 lint: venv
 	pylint ./src
