--- src/constants/Enums.py
@@ -1,4 +1,0 @@
-import enum
-
-class Tables (enum.Enum):
-    users = 'USERS_TABLE'
